refactor: split launch_simulation into subclasses
